refactor: organize inbox components into subpackage
