+++ feature/inbox/src/main/java/com/shrivatsav/monomail/feature/inbox/components/SwitchAccountCard.kt
@@ -1,4 +1,6 @@
-package com.shrivatsav.monomail.feature.inbox
+package com.shrivatsav.monomail.feature.inbox.components
+
+import com.shrivatsav.monomail.feature.inbox.*
 
 import com.shrivatsav.monomail.ui.components.AvatarCircle
